2 new samples for monitoring alert api.
In the new canonical snippet format.
