+++ monitoring/api/AlertTestFixture/AlertTestFixture.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Google.Cloud.Monitoring.V3" Version="1.1.0-beta01" />
+  </ItemGroup>
+
+</Project>
